More useful error message if proxy is not initialized.
This commit also makes the proxy check a tool.
